+++ packages/integration-sdk-cli/src/commands/collect.ts
@@ -9,7 +9,7 @@ import * as log from '../log';
 import { loadConfig } from '../config';
 
 // coercion function to collect multiple values for a flag
-const collecter = (value: string, arr: string[]) => {
+const collector = (value: string, arr: string[]) => {
   arr.push(...value.split(','));
   return arr;
 };
@@ -22,13 +22,17 @@ export function collect() {
     .option(
       '-s, --step <steps>',
       'step(s) to run, comma separated if multiple',
-      collecter,
+      collector,
       [],
     )
+    .option('-V, --disable-schema-validation', 'Disable schema validation')
     .action(async (options) => {
+      const enableSchemaValidation = !options.disableSchemaValidation;
       const config = prepareLocalStepCollection(await loadConfig(), options);
       log.info('\nConfiguration loaded! Running integration...\n');
-      const results = await executeIntegrationLocally(config);
+      const results = await executeIntegrationLocally(config, {
+        enableSchemaValidation,
+      });
       log.displayExecutionResults(results);
     });
 }